restructure fro detector , captor
