+++ cfg/coco.data
@@ -2,6 +2,6 @@ classes= 80
 train  = E:/MSCOCO/trainvalno5k.txt
 #train = E:/MSCOCO/5k.txt
 valid  = E:/MSCOCO/5k.txt
-names = C:/Users/iApp/Desktop/projects/cardetection/data/coco.names
+names = C:/Users/iApp/Desktop/projects/darknet_detector/data/coco.names
 backup = backup
 eval=coco
